kvx: eth: add tx_fcs_offload sysfs entry
Summary:
For testing/debug purpose, we want to be able to disable (set to 0) the
add_fcs_en attribute on kvx MAC lanes.

When enabled (which is the default), the MAC will insert the FCS 4 bytes
CRC at the end of the ethernet packet.

This option is a simple boolean 0 or 1, disabling to enabling the all 4
lanes.

Test Plan: ODP

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: tcostis, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D6579