remoteproc: core: unregister virtio device after stop
Summary:
Currently, the virtio device are unregistered before stopping the rproc
device. THis is a potential cause of error since if an interrupt
triggers between these two calls, then virtio rings might receive an
interrupt via vring_interrupt which will end up on a freed vring.
In order to fix this, move stop function of subdev to unprepare. The
effect is that virtio device are unregistered after stopping the remote
processor and thus interrupts are disabled.

Credits goes to @vchardon for fixing it on PCIe driver

Test Plan: Run OpenCL testsuite multiple times.

Reviewers: vchardon, O51 Linux Coolidge

Reviewed By: vchardon, O51 Linux Coolidge

Subscribers: vchardon, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D6599