kvx: add functions to writeback/invalidate dcache
Summary:
We are going to add a syscall to flush dcache from userspace. to do so,
we need to add a function to allow flushing cache using addresses from
userspace. This can be done by walking the page table to find the match
between the the virt -> phys.

Ref T10801

Depends On D6322

Test Plan: Add a test in coolval which uses next to come cachectl syscall

Reviewers: O51 Linux Coolidge, gthouvenin, jmaselbas

Reviewed By: O51 Linux Coolidge, gthouvenin, jmaselbas

Subscribers: jmaselbas, gthouvenin, #linux_coolidge_cc

Maniphest Tasks: T10801

Differential Revision: https://phab.kalray.eu/D6324