iommu: kvx: Fix iova to physical address calculation for hugepages
Summary:
kvx_iommu_iova_to_phys() was not able to find the physical address of
an iova address if the address was bigger than the page address + 4KB,
for instance for hugepages. The PN TLB entry value was calculated assuming 4KB
pages.

This patch fix this, the PN TLB entry value is adapted for each pagesize
during the lookup.

Test Plan: Run SPDK app with NVME PCI driver which uses 2M hugepages

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D6915