kvx: set access flags and wprotect for contiguous huge page
Summary:
We need to set the dirty and write flags for all contiguous huge
pages that belongs to the same physical TLB for coherency. When flags
are set entries are now flushed and all entries are synchronized.

Ref T13802

Test Plan: Run hugemap test in a loop and check that it is running smoothly

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13802

Differential Revision: https://phab.kalray.eu/D7170