remoteproc: kvx: do not read cluster status in probe
Summary:
When booting from flash, the FSBL leaves the cluster running but in
stop mode. When starting Linux, the driver believes the cluster is
running and will refuse to start anything. Since we want to be able to
use the cluster, simply reset it.

Ref T13833

Test Plan: Run remoteproc test after booting from flash

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13833

Differential Revision: https://phab.kalray.eu/D7228