kvx: net: hard limit on some Rx parser configurations
Summary:
Rx parsers can extract the index of the current layer and for some
layers, check the checksum. This information is forwarded through the Rx
metadata.

This Rx metadata embeds at most 4 extracted indexes and 4 CRC errors.
If a parser tries to configure more than 4 extract index or 4 CRC check,
it will work, but the hardware is not well used.

The user has no control at this grain: the parsers are decided at higher
level, this configuration is done at module level.

To avoid any misunderstanding, the module will fail with a BUG_ON if a
parser exceed this limit.

Test Plan: NA

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D7555