kvx: rework put/get_user to fix bug and update
Summary:
p parameter was evaluated more than once leading to double increment if
called with sp++ for instance.
This was the case in create_elf_table and this resulted in a corrupted
stackframe for the userspace.
Fix this by using an intermediate variable (__p) to avoid evaluating
twice the parameters.
Moreover, the __err_ and __error macros are not used anymore and can be
removed.

Ref T13785

Test Plan: Run kernel 5.8

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13785

Differential Revision: https://phab.kalray.eu/D7648