kvx: eth: define lane id from device tree
Summary:
The lane id depend on signal routing on the PCB and the configuration.
In particular when some lanes are aggregated the lane number are no
longer consecutive. Thus in order to support all case, allow the lane
number to be explicitly specified in device tree.
For backward compatibility the previous lane number allocation has been
preserved, and will be used if the lane number is not defined in the
device tree.

Test Plan: Confirm that ethernet is still working

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: bmugnier, tcostis, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D7422