kvx: eth: properly set pre and post parameter in phy
Summary:
The offset was not computed correctly and the values used to be written
in a wrong registers. Properly compute address from the beginning
without using the previous value.

Test Plan:
Change the parameters in device tree with kalray,phy-param entry
Confirm with an oscilloscope that pre and post values are changed

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D8250