kvx: net: extend sysfs with parser information
Summary:
In order to use hardware acceleration for Rx packets, we need to have a
precise description of what is enabled for each parser.

This patch aims to expose the whole configuration for each parsers, in a
dedicated `parser/` subdirectory. It looks like this:

/sys/class/net/enmppaX/
               └── parser/
                   ├── 0/
                   │   ├── enable
                   │   └── rule/
                   │       ├── 0/
                   │       │   ├── enable
                   │       │   ├── type
                   │       │   ├── add_metadata_index
                   │       │   └── check_header_checksum
                   │       ├── ...
                   │       └── 3/
                   ├── ...
                   └── 31/

fix T13927

Test Plan: used in dcbu

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T13927

Differential Revision: https://phab.kalray.eu/D8294