kvx: ftrace: Add support for function tracer
Summary:
Add support for HAVE_FUNCTION_TRACER and HAVE_FUNCTION_GRAPH_TRACER.

The mcount symbol is __mcount. This patch adds the __mcount function
that The mcount function should check the function pointer
ftrace_trace_function to see if it is set to ftrace_stub. If it is,
there is nothing to do and just return immediately. If it isn't,
then call that function in the same way the mcount function normally
calls __mcount_internal:
  - the first argument is the "frompc"
  - the second argument is the "selfpc" (adjusted to remove the
    size of the mcount call that is embedded in the function).

To also work for function graph tracer it modified
the __mcount function to check the function pointers
ftrace_graph_return (compare to ftrace_stub) and
ftrace_graph_entry (compare to ftrace_graph_entry_stub).
If either of those is not set to the relevant stub
function, call the arch-specific function ftrace_graph_caller
which in turn calls the arch-specific function prepare_ftrace_return.

Ref T13869
Ref T12412

Test Plan:
To check that it is working we
  - mount debugfs: mount -t debugfs none /sys/kernel/debug/
  - go into tracing dir:  cd /sys/kernel/debug/tracing/
  - start tracing function: echo function > current_tracer
  - wait a little bit do some stuff
  - dump trace: cat trace > /tmp/trace.txt
  - stop tracing: echo nop > current_tracer
  - check that the trace is correct.
Check again but with function_graph. It is the same test as above but
you echo "function_graph" instead of "function"

Reviewers: O51 Linux Coolidge, cleger

Reviewed By: O51 Linux Coolidge, cleger

Subscribers: cleger, tcostis, #linux_coolidge_cc

Maniphest Tasks: T12412, T13869

Differential Revision: https://phab.kalray.eu/D4939