irqchip: kvx-gic: Enable parent IRQ for all CPUs
Summary:
When routing GIC inputs to other CPUs, the parent IRQ on the core
interrupt controller must be enable. This can be done using a startup
function for other CPUs. This allows to enable the line on all CPUs and
not only on CPU 0.

Ref T9383

Test Plan: Try to switch IRQs to a new cpu and check it works with this patch

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9383

Differential Revision: https://phab.kalray.eu/D9621