kvx: enable sparsemem support
Summary:
Currently, DDR is configured in banked mode which means that the RAM is
split in two sections of 4G with a huge hole between them. Tihis is not
optimal using a flatmem [1] model leading to a lot of memory lost for this
hole. This patch enables sparsemem which allows to reduce the mem_map
size by using sections.

[1] https://www.kernel.org/doc/html/latest/vm/memory-model.html

Test Plan: Compile and run

Reviewers: O51 Linux Coolidge, jmaselbas, gthouvenin

Reviewed By: O51 Linux Coolidge, jmaselbas, gthouvenin

Subscribers: gthouvenin, jmaselbas, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D9362