kvx: mm: do not raise error when refilling device memory
Summary:
Previously when a refill is performed on a mmu page fault a check is
performed in order to determine if the pfn is valid. This check is done
with pfn_valid function, but pfn_valid return true only when the pfn
correspond to a memory location (as it should). Thus the refill will
fail when a device memory (mapped using ioremap function) is being added
in the TLB. In the kvx_mmu_jtlb_add_entry adding tlb for device address
is valid, thus remove the test.

Test Plan:
Make test on Winstron platform, with some stress on NVME device in order
to force pcie location TLB be evicted and reloaded at some time.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D9833