irqchip: kvx-pcimsi: add proper set_affinity to chip
Summary:
In order to set affinity properly, we need to call set_affinity on the
parent. Indeed, the chained handler breaks the irq hierarchy and thus,
there is not real parent associated to the final irq_data. In order to
properly set the affinity on the parent, call irq_set_affinity using the
parent_irq value.

Ref T9383

Depends on D9623

Test Plan:
Change e1000 irq affinity on another cpu and check it correctly switch
to the new cpu.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9383

Differential Revision: https://phab.kalray.eu/D10077