irq: add irq_chip_write_msi_msg_done callback
Summary:
In the case of our mailbox IRQ controller, we need to have the
confirmation that the device updated its MSI address. The only way to
have that is to add such callback. This callback is called on the MSI
provider once the irq_chip_write_msi_msg has been called on the device.

Ref T9383

Test Plan: Switch affinity and check that kvx_mailbox_msi_write_msg_done is called

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9383

Differential Revision: https://phab.kalray.eu/D10229