kvx: get rid of all seg/get_fs stuff
Summary:
set_fs,get_fs provided protection for kernel access to userspace. This
was a soft mecanism to forbid kernel to access userspace in normal time
but could be enable at special places. This has now been removed and
we can implement it on kvx using a specific ASN for the kernel.
Currently, on kvx, the kernel can always access the user space and
this can be a source of exploit.
enable_user_access and disable_user_access are added as stubs and will
be implemented later to do what they say using a specific ASN.

Test Plan: Use 5.10 rebase and boot

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D10314