spi: dw: add kvx driver for DW ssi controller
Summary:
This is not a from scratch commit but only a port of previous spimem
driver for the new framework. Thanks to the possibility of adding
new device with differen memops than the standard one, we can add
this driver without touching the generic code.

Test Plan: Execute the flash tests

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D10386