registers: add kvx-ftu's cluster status definition
Summary:
The FTU contains a register named cluster status that was missing some
fields in its definition.

This patch adds them, and in particular the RST bit, image of the
cluster's reset, useful when making sure the cluster is ready to accept
AXI transactions.

Test Plan:
Next commit adds a check in remoteproc in order to wait for the end of a
cluster reset before loading anything in its memory.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D11614