kvx: eth: 2 parser groups + checksum capability
Summary:
The second group is a clone of the first one hardware wise. In software we
only save the top half.
We also make sure the user doesn't want to use the unicast bit for mac
address as it is required for this to work.
Parsers are grouped by checksum capability, either no checksum
computation possible, one, or four.
Driver handles an indirection table so there is no need for the user
to know which parser handles which checksum capability.

Parser sysfs represent the hardware view: it show the 32 physical
parsers, not the 16 indirect ones.

Additional debug info for parser setup.

Fix T15825

Test Plan: MF

Reviewers: O39 ethernet, bmugnier, tcostis

Reviewed By: O39 ethernet, tcostis

Subscribers: ysionneau, #linux_coolidge_cc, jreybert, tcostis

Maniphest Tasks: T15825

Differential Revision: https://phab.kalray.eu/D11949