kvx: allow L2 cache being disabled in SMP when QEMU is detected
Summary:
When running under QEMU, the L2 cache controller is not strictly
necessary for correct operations, even in SMP. This is because caches
are not simulated in QEMU, implying a coherent memory view for all CPUs
at any given time.

This commit adds a way of detecting we're running under QEMU, with the
"kalray,is-qemu" property of the L2 cache node in the DTB. When present
and when the L2 cache node is explicitly disabled, the kernel will
continue booting normally even with CONFIG_SMP on.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D13930