kvx: core timer: set minimum delta value to 1 tick
Summary:
The KVX core timer is programmed to trigger in n ticks by setting its
value register to n - 1. This is because the actual IRQ is raised when
the timer underflows (from 0 to UINT64_MAX).

A minimum delta ticks value of 0 is wrong because it leads to setting
the value register to UINT64_MAX. This commit sets it to 1.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D14226