Early watchdog enable and ack to synchronize with L2 firmware
Summary:
For robust watchdog protection in both
Linux and L2 firmware communication at boot time,
we need to add an acknowledge from Linux to the L2
firmware and enable watchdog early in Linux boot stage.
For the L2 synchronizetion, we use a new fields in
the status register of the L2 firmware to make it possible.

Ref T17455

Test Plan: Linux, CI, L2

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D16366