kvx: remoteproc: Prevent Linux from resetting active cluster or cluster0
Summary:
The FTU can't reset Cluster0, so throw an error if something tries to do
this via `kvx_rproc_reset`. Also, make sure that even if a cluster is in
the list of remoteprocs don't reset it if its the active cluster
executing Linux. E.g., if someone starts Linux on, Cluster1/2/3/4 with
the default DTB file the system will crash because Linux will reset
itself. This patch prevents this!

Ref T17738

Test Plan: tbd

Reviewers: ysionneau, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T17738

Differential Revision: https://phab.kalray.eu/D17989