kvx: Removed spinlock in context where Linux might call sleep
Summary:
The access to the re-timers is always bundled in two i2c operations.
First, the lane (that should be manipulated) is selected and then the
write is performed to set the parameters for that given lane. So, each of
these bundles must be wrapped in a spinlock. However, for this specific
case the `retimer_cfg` is only called in the `ti_rtm_probe` function, which
is only entered a single time on a single core. So, no need to have a
spinlock here.
We decided to remove it because when compiling the `conf_mppa_debug`
Linux complains about `BUG: scheduling while`atomic`, because the
platform independant `i2c_write` can actually be interrupted.

Fixes f63a0f88be74 ("ti: retimer: i2c access in critical section")

Ref T18181

Test Plan:
Jenkins Pipeline, and make sure that when running
conf_mppa_debug that Linux doesn't complain about `BUG: scheduling while
atomic`

Reviewers: ysionneau, tcostis, O39 ethernet!

Maniphest Tasks: T18181

Differential Revision: https://phab.kalray.eu/D18130