kvx: manipulating the ASN in the mm_struct should be done with interrupts off
Summary:
If the number of mappings that should be removed from the TLB exceeds the size
of the TLB it's faster to switch to a new ASN instead of shooting down
individual pages in the TLB. Then next time this context is activated the
function `get_mmu_context` assigns a new ASN to it. But setting the ASN in the
mm struct to `MM_CTXT_NO_ASID` should be done with interrupts off
(`local_irq_save`) to prevent weird behaviour or the usage of old already
invalid pages.
Also if an other PE requests a remote TLB shootdown the function
`local_flush_tlb_page` is called to flush the page locally. But again this
function might get interrupted between obtaining the ASN from the current mm
struct and the actual hardware flush operation. In the meantime the ASN might
have been changed or invalidated which in turn might cause weird behaviour. So,
IRQs should be disabled while doing the entire "look-up + flush".

Ref T18311

Test Plan: CI

Reviewers: ysionneau, jmaselbas, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T18311

Differential Revision: https://phab.kalray.eu/D19253