smp: Fix to take into consideration nr_cpus on Linux command line
Summary:
Starting Linux with `nr_cpus=X` doesn't have any effect on the nr of CPU that
can be used or that are online. It will just start with all cores that are
defined in the `.config` (using `CONFIG_NR_CPUS`). With this fix it correctly
limits the nr of CPUs (`num_possible_cpus`) to the number specified via
`nr_cpus` on the command line.

Ref T18874

Test Plan: Successful pipeline

Reviewers: jmaselbas, ysionneau, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T18874

Differential Revision: https://phab.kalray.eu/D19826