kvx: pgtable: fix pgprot_noncached to really make prot uncached
Summary:
Fix pgprot_noncached to really make prot uncached.

If the prot was for instance already PAGE_KERNEL, meaning it
already had been or'ed with _PAGE_CACHED, or'ing it with
_PAGE_UNCACHED (which is 0) won't clear the bits that
were set to 1 because of _PAGE_CACHED.

The cache policy bits need to first be cleared and then
OR'ed with the correct value.

On CV2 _PAGE_CACHED is 0111, and _PAGE_UNCACHED is 0000
meaning that _PAGE_CACHED | _PAGE_UNCACHED ends up
still being 0111 which is "I/D cached".

On CV1 _PAGE_CACHED is 10, and _PAGE_UNCACHED is 01
meaning that _PAGE_CACHED | _PAGE_UNCACHED ends up
being 11 which means Data uncached and Insn cached.
This was fine most of the time, except if a coherent
mapping was created to write instructions to it.

Ref T18901

Test Plan: CI

Reviewers: O51 Linux Coolidge, jmaselbas

Reviewed By: O51 Linux Coolidge, jmaselbas

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T18901

Differential Revision: https://phab.kalray.eu/D20127