kvx: fixed distinction between swap entry and non-present
Summary:
PTE entries that are not present are fetched from a swap device and Linux
identifies them with the function `pte_present` (which checks that the
`_PAGE_PRESENT` bit[0] is NOT set . But PTE entries with `PROT_NONE`
(`PAGE_NONE`) are also just all zeroes. So, Linux can't understand the
difference between a swapped out page and a page with `PROT_NONE`. To fix this
a `PROT_NONE` page now has bit[1] (`_PAGE_ACCESSED` -> kv3-1 or `_PAGE_HUGE`
-> kv3-2) set to distinguish the two.

Ref T18766

Test Plan: Successful pipeline

Reviewers: ysionneau, O51 Linux Coolidge!

Subscribers: ysionneau, jmaselbas, jvetter, #linux_coolidge_cc

Maniphest Tasks: T18766

Differential Revision: https://phab.kalray.eu/D20313