kvx: eth: fix uninitialized id field in struct kvx_eth_dt_f
Summary:
The id field was left uninitialized in the dispatch table structure.
Since the struct is allocated with devm_kzalloc, the field was 0,
leading to all the entries effectively pointing to the hardware entry 0.

Fix this by splitting initialization from reset. Set the id field during
initialization, and the other fields during reset as before.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: tcostis, ysionneau, O39 ethernet!

Differential Revision: https://phab.kalray.eu/D20987