irq: add irq_chip_write_msi_msg_done callback
In the case of kvx mailbox IRQ controller, we need to have the
confirmation that the device updated its MSI address. The only way to
have that is to add such callback. This callback is called on the MSI
provider once the irq_chip_write_msi_msg has been called on the device.

Signed-off-by: Yann Sionneau <ysionneau@kalray.eu>