arch: kvx: smp: ensure PE boot address is correct
Summary:
When kicking the PE for SMP boot, they were simply woken up through the
power controller.
While it has worked so far since the default value of their reset PC
happened to be aligned to kvx_start, this assertion proved to be false
when tinkering with lower-level bootloaders.

This patch forces the cluster-local reset_pc for the PE in order to make
sure booting them will end up in Linux and nowhere else.

Ref T19451

Test Plan:
play with the reset_pc register in lower levels of the bootchain and try
to boot a SMP Linux image.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D21654