phy: Convert usb-nop-xceiv to the generic phy subsystem
Summary:
I recently encountered some limitations when using the usb-nop-xceiv
driver using the usb-phy subsystem. The situation is: I have a board
with two external usb phys (ULPI) that only needs to be reset by the
USB controller. However both phys shares the same reset line, which
will get toggled twice by the USB driver, once by controller.

This is what I observed with the current Linux usb-nop-xceiv driver:
          _______                            ___      _________
   RESET:        \__________________________/   \____/
                1^                         2^  3^   4^
 1. and 2. first controller toggle the reset
 3. and 4. second controller toggle the reset

After the second toggle the first driver didn't expected the reset and
causing issues, but I don't recall what exactly, it is maybe a driver
that simply fail to get probed.

The following is what I observed when using the "usb-nop-xceiv"
driver in Barebox, which is a bit different than the Linux one.

           _______                            _________________
    RESET:        \__________________________/
                 1^                         2^

In this case the reset is only toggled once, this is because the phy
driver count the number of time it has called init and power_on, which
is exactly what is done by the generic phy subsystem in Linux, but not
by the "legacy" usb-nop-xceiv usb-phy driver.

This difference in behavior is what motivated to "port" the existing
usb-nop-xceiv driver to the generic phy subsystem. However this is far
from being complete... I've only tested the reset function for this nop
phy, and it lack a lot of features from the original driver to be a
valid replacement for it.

I would like to know if there is an interest in converting the current
usb-nop-xceiv driver to the generic phy subsystem?

I would like to keep the usb-nop-xceiv compatible as it used by Barebox
and it works OK. If a new compatible is added then device-tree nodes
shouldn't have both, the new and usb-nop-xceiv, compatible because of
the specific check in phy-core.c which will always return ENODEV.

Test Plan: to be tested on a k200 rev2

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T16089

Differential Revision: https://phab.kalray.eu/D11851