eth: qsfp: check capability bit before tuning qsfp params
Summary:
Check the capbility bit of each qsfp param before writing it on eeprom. If
a feature is not implemented, the param is not tuned.
Additionnally, the capability bit might be inaccurate for some cables. In
this case, it the bit is 0, we try to write it anyway. In case of failure,
only "safe" qsfp params are tuned.

Ref T19736 T18798

Test Plan: CI + acceptance

Reviewers: O39 ethernet, tcostis

Subscribers: tcostis, gpommeret

Maniphest Tasks: T19736

Differential Revision: https://phab.kalray.eu/D22567