kvx: iommu: Fix use-after-free with kvx_domain
Summary:
If we let a kvx_domain in a list and we free it, adding a new element
after the list will override the "prev" or "next" field in the freed
kvx_domain, causing a use-after-free.
Ref T19862

Test Plan: Manually verified that issue was gone with debug options activated in kernel.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T19862

Differential Revision: https://phab.kalray.eu/D23241