dma: kvx: fix dma comp queue full
Summary:
The DMA completion queue full
notification address of the RX chan
was not initialized (ie: 0).
Indeed, if the DMA RX chan gets full, this address
will be used by the DMA and an IOMMU error may
trig when the DMA tries to access the 0 IOVA.
We fix this by using the same notification
address of an RX completion.

Test Plan: CI

Reviewers: O39 ethernet!

Differential Revision: https://phab.kalray.eu/D23322