kvx: Added generic ASLR support
Summary:
For all architectures a generic mechanism was introduced to enable ASLR for
user space mmap calls. This changes the memory layout of user space
processes to a top down layout. This means mmap'ed memory grows directly
from below the stack towards smaller addresses (like the stack) as opposed to
before, where it was growing upwards towards the stack. Also the base
address of this memory is now randomized with a configurable number of
bits used for randomization. The full patchset for the Linux kernel can be
found here: https://lkml.org/lkml/2019/3/22/136). This commit enables the
necessary Kconfig flags for kvx to support this generic mechanism.

Ref T10524

Test Plan: CI

Reviewers: ysionneau, jmaselbas, O51 Linux Coolidge

Reviewed By: ysionneau, O51 Linux Coolidge

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T10524

Differential Revision: https://phab.kalray.eu/D23430