kvx: remove rm_firmware launch from Linux
Summary:
The rm_firmware used to be launched directly by Linux when spawned from
the RM; but recent work has led to deploy it directly from the
bootloader, and spawn the Linux ELF directly on a PE instead.

This patch removes support for the elf-embedded rm_firmware, and all
code pertaining to it.
The assumption is now that the processor booting the code should be able
to directly run Linux. If RM services are described in the DeviceTree,
their RM counterpart should have been launched by the lower levels on
the bootchain.

Ref T20630

Test Plan:
CI: L2 sync should work both when launched with jtag-runner as with boot
from flash

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20630

Differential Revision: https://phab.kalray.eu/D26269