kvx: Remove option to disable the Secure Dame handling
Summary:
There is a bug in `entry.S`, where `$ilr` is saved if the secure dame handling
config option is enabled. But in the retore path `$ilr` is unconditionally
restored. So, if the config option is disabled `$ilr` will be restored
with garbage. Moreover, the system is never tested without this option
and opens a big breach between kernel and user. So, why should we
disable it, or have it in the first place?

Ref T20648

Test Plan: CI

Reviewers: ysionneau, jmaselbas, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20648

Differential Revision: https://phab.kalray.eu/D26426