kvx: eth: cv2 - udpate for lb RFS handling: sysfs, irq
Summary:
We update the content of the kvx-net sysfs in order to allow
simple access to the EthRx RFS block of the load balancer.
This includes:
  param_fk_cmd: set to 1 for writing the fk_part @ fk_idx
  ctrl_hash_rss_ena: same as register bit
  ctrl_rfs_ena: same as register bit
  seed_command: 0 to write the part in hash0 seed @row/idx
                1 to write the part in hash1 seed @row/idx
  fk_command: same as bitfield in register
  counter of RFS table corruption counter
Also added, handling of the Irq ethernet rx_error with a
simple counter increment.

Ref T20054

Depends On D27200
Depends On D27199
Depends On D27198
Depends On D24240
Depends On D24142
Depends On D24139
Depends On D23978
Depends On D23774

Test Plan: CI: CV1 testsuite, CV2 validsuite (partial)

Reviewers: O39 ethernet!, O129 linux drivers

Maniphest Tasks: T20054

Differential Revision: https://phab.kalray.eu/D27201