kvx: Replaced handcrafted asm syscall entry with C function
Summary:
Our syscall handler is fully written in asm, but this won't work well in
the future with the new generic entry/exit functions that streamline
tracing, and scheduling work in the kernel. So, replace the asm with
some small boilerplate code to then call a C function similar to what is
done on risc-v.

Ref T20648

Test Plan: CI

Reviewers: ysionneau, jmaselbas, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20648

Differential Revision: https://phab.kalray.eu/D26455