kvx: entry: Consolidate ret_from_kernel_thread into ret_from_fork
Summary:
The ret_from_kernel_thread() does the same as ret_from_fork(), the only
difference is whether to call the fn(arg) or not, this can be achieved by
testing fn is NULL or not, i.e $r20 is 0 or not. Many architectures have done
the same thing, it makes entry.S more clean.

Ref T20648

Test PlanL: CI

Reviewers:

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20648

Differential Revision: https://phab.kalray.eu/D26458