kvx eth: link status from retimer instead of phy
Summary:
In the case of fiber cable, link status is handled by a qsfp interrupt, when the cdr lock is lost.
In the case of copper cable, link status is checked at PHY level, with linklos register.
This commit harmonizes the kvx eth driver and thus, in the case of copper cable, cdr lock from retimer is checked instead of linklos.

Ref T21248

Depends on D28700

Test Plan: CI eth

Reviewers: O129 linux drivers

Maniphest Tasks: T21248

Differential Revision: https://phab.kalray.eu/D28701