kvx: dame: Don't check DAME errors in syscall
Summary:
DAME errors are expected when trying to talk to a PCI device that is
removed. This shouldn't trigger a kernel panic.
With this patch, we simply go back to the old behavior before
48ddd23dc9b8, where DAME were not checked in syscalls.
However, I'm wondering if we should not remove the kernel panic for all
cases of DAME errors.

Test Plan: Dryrun OK

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D28981