kvx-net: find autoneg common speed earlier
Summary:
In the current implementation, the negociated abilities, including speed and fec, are retrieved from the BP_ETH_STATUS registers in mac_ctrl.
Those registers are asserted only when the entire autoneg sequence is done and the state AN_GOOD_CHECK is reached, meaning at the very end.
In order to save some time, it is possible to negociate the abilities earlier right after the base page exchange, by comparing manually the
local and remote device abilities. This would allow to enable link training and start to configure phy/pcs before the end of the autoneg.

This commit implements the following:
- firstly, in the base page exchange state, wait for the page received bit assertion, whhich means that base page exchange is done (this bit
  was previously never asserted for unknown reasons)
- secondly, read the local device and link partner abilities in order to compare them. Select the highest speed available.

Ref T20034

Test Plan: tests eth autoneg

Reviewers: O129 linux drivers

Maniphest Tasks: T20034

Differential Revision: https://phab.kalray.eu/D29278