kvx-eth: setup the link even without qsfp
Summary:
The kvx_eth_setup_link call is surrounded with a check on whether we
have QSFP or not. This prevents the netdev to come up in the no QSFP
case.

This check is justified when we're runnig on HAPS because the link is
already setup by another mean. However in other cases where we don't
have a QSFP (such as QEMU), we still want the netdev to go up.

Replace this check with a call to kvx_eth_is_haps to discriminate
between those two cases.

Fixes: c751d5baf3bde1c8c03d156baf98ddc2d6d68105

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O129 linux drivers

Maniphest Tasks: T21482

Differential Revision: https://phab.kalray.eu/D29462