kvx eth: call autoneg FSM even if autoneg is disabled
Summary:
Whether the autoneg is enabled or not, the mac/phy/pcs configuration remains the same. The
only difference is the autonegociation itself and the link training.
This commit simplifies link configuration so that the autoneg FSM is systematically called,
whether the autoneg is enabled or not. Transitions were added to skip autoneg and link training
when autonegociation is disabled. The function kvx_eth_an_execute() was renamed for clarity.
In order to allow kvx_eth_mac_pcs_pma_hcd_setup() splitting/reworking, its content is moved the
state AN_STATE_PCS_CFG of the autoneg FSM.

Ref T21766

Test Plan: CI

Reviewers: O129 linux drivers

Maniphest Tasks: T21766

Differential Revision: https://phab.kalray.eu/D30452