kvx eth: feedback pcs status to AN module
Summary:
AN and LT ref guide specifies that once the autonegociation and link training are done, the
PCS link status shall be synchronized with the AN module. Additionally, Fig. 1 shows that the
link status is generated from align_done, block_lock and hi_ber.
In the PCS link status register, only the bit of the lane used for autonegociation is set.

Ref T21965

Test Plan: CI

Reviewers: O129 linux drivers

Maniphest Tasks: T21965

Differential Revision: https://phab.kalray.eu/D31452