kvx eth: set correct training pattern on each lane
Summary:
The training pattern parameters are not configured prior to link training, thus some default
parameters are used for all lanes. IEEE 802.3 clause 92.8 specifies that distinct polynomials
and seeds shall be set for each physical lane in the case of 100G, in order for the training
pattern checker of the link partner to be able to check properly each packet.
This commit sets the PRBS parameters specific to each lane defined in Table 92-5 of the
standard, before enabling the link training. When the negotiated speed is not 100G, default
values are set.

Ref T21002 T20034 T21865

Test Plan: CI

Reviewers: O129 linux drivers

Maniphest Tasks: T21865, T21002, T20034

Differential Revision: https://phab.kalray.eu/D31822