kvx: Clear i/d cache before enabling them
Summary:
We make sure the i/d caches are cleaned before enabling them. Also, we
make sure there are no outstanding loads and/or stores before enabling
the caches. A problem related to this was observed in Barebox, so do the
same measures in Linux to be sure we don't have the same problem here.

Ref T21986

Test Plan: CI

Reviewers: O51 Linux Coolidge, ysionneau

Reviewed By: O51 Linux Coolidge, ysionneau

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T21986

Differential Revision: https://phab.kalray.eu/D31570